split benchmark
