+++ code/source/unittest/commands.c
@@ -11,7 +11,7 @@ Description:
 ==============================================================================
 */
 #include "fossil/unittest/commands.h"
-#include "fossil/unittest/internal.h"
+#include "fossil/unittest/framework.h"
 #include "fossil/unittest/console.h"
 #include <stdio.h>
 #include <stdlib.h>
